fix: make react-native-image-crop-picker optional (#2603)
* fix: make react-native-image-crop-picker optional

* fix: tests

* docs: add optional deps docs

* docs: add note for the optional dependency
