+++ docusaurus/docs/reactnative/common-content/ui-components/channel/props/has_camera_picker.mdx
