+++ docusaurus/docs/reactnative/common-content/ui-components/channel/props/has_image_picker.mdx
@@ -2,4 +2,4 @@ Enable the image picker on the [`MessageInput`](../../../../ui-components/messag
 
 | Type    | Default |
 | ------- | ------- |
-| boolean | true    |
+| Boolean | `true`  |
