+++ docusaurus/docs/reactnative/core-components/channel.mdx
@@ -74,6 +74,7 @@ import HandleQuotedReply from '../common-content/ui-components/channel/props/han
 import HandleReaction from '../common-content/ui-components/channel/props/handle_reaction.mdx';
 import HandleRetry from '../common-content/ui-components/channel/props/handle_retry.mdx';
 import HandleThreadReply from '../common-content/ui-components/channel/props/handle_thread_reply.mdx';
+import HasCameraPicker from '../common-content/ui-components/channel/props/has_camera_picker.mdx';
 import HasCommands from '../common-content/ui-components/channel/props/has_commands.mdx';
 import HasFilePicker from '../common-content/ui-components/channel/props/has_file_picker.mdx';
 import HasImagePicker from '../common-content/ui-components/channel/props/has_image_picker.mdx';
@@ -523,6 +524,10 @@ The max allowable is 255, which when reached displays as `255+`.
 ### handleThreadReply
 
 <HandleThreadReply />
+
+### `hasCameraPicker`
+
+<HasCameraPicker />
 
 ### `hasCommands`
