+++ examples/SampleApp/package.json
@@ -42,7 +42,7 @@
     "react-native-fs": "^2.18.0",
     "react-native-gesture-handler": "^2.14.0",
     "react-native-haptic-feedback": "2.0.3",
-    "react-native-image-crop-picker": "0.39.0",
+    "react-native-image-crop-picker": "^0.41.2",
     "react-native-image-resizer": "1.4.5",
     "react-native-markdown-package": "1.8.2",
     "react-native-quick-sqlite": "8.0.2",
