+++ package/expo-package/src/index.js
@@ -13,7 +13,6 @@ import {
   oniOS14GalleryLibrarySelectionChange,
   saveFile,
   Sound,
-  takePhoto,
   Video,
 } from './handlers';
 
@@ -21,6 +20,7 @@ import {
   pickDocument,
   setClipboardString,
   shareImage,
+  takePhoto,
   triggerHaptic,
 } from './optionalDependencies';
