+++ package/expo-package/src/optionalDependencies/index.ts
@@ -3,3 +3,4 @@ export * from './shareImage';
 export * from './pickDocument';
 export * from './triggerHaptic';
 export * from './Video';
+export * from './takePhoto';
