+++ package/native-package/package.json
@@ -15,17 +15,17 @@
   },
   "peerDependencies": {
     "@react-native-camera-roll/camera-roll": ">=5.0.0",
-    "@react-native-community/netinfo": ">=2.0.7",
     "@react-native-clipboard/clipboard": "^1.11.1",
+    "@react-native-community/netinfo": ">=2.0.7",
     "@stream-io/flat-list-mvcp": "^0.10.3",
     "react-native": ">=0.60.0",
+    "react-native-audio-recorder-player": ">=3.6.4",
     "react-native-document-picker": ">=9.0.1",
     "react-native-fs": ">=2.16.6",
     "react-native-haptic-feedback": ">=1.11.0",
     "react-native-image-crop-picker": ">=0.33.2",
     "react-native-image-resizer": ">=1.4.2",
     "react-native-share": ">=4.1.0",
-    "react-native-audio-recorder-player": ">=3.6.4",
     "react-native-video": ">=6.4.2"
   },
   "peerDependenciesMeta": {
@@ -39,6 +39,9 @@
       "optional": true
     },
     "react-native-haptic-feedback": {
+      "optional": true
+    },
+    "react-native-image-crop-picker": {
       "optional": true
     },
     "react-native-audio-recorder-player": {
@@ -58,7 +61,6 @@
     "@stream-io/flat-list-mvcp": "0.10.3",
     "react-native": ">=0.60.0",
     "react-native-fs": ">=2.16.6",
-    "react-native-image-crop-picker": "^0.38.0",
     "react-native-image-resizer": ">=1.4.2"
   }
 }
