+++ package/native-package/src/handlers/index.ts
@@ -4,7 +4,6 @@ export * from './getLocalAssetUri';
 export * from './getPhotos';
 export * from './NetInfo';
 export * from './saveFile';
-export * from './takePhoto';
 export * from './Sound';
 export * from './Video';
 export * from './oniOS14GalleryLibrarySelectionChange';
