+++ package/src/components/Channel/Channel.tsx
@@ -75,7 +75,7 @@ import {
   ThumbsUpReaction,
   WutReaction,
 } from '../../icons';
-import { FlatList as FlatListDefault, pickDocument } from '../../native';
+import { FlatList as FlatListDefault, isImagePickerAvailable, pickDocument } from '../../native';
 import * as dbApi from '../../store/apis';
 import type { DefaultStreamChatGenerics } from '../../types/types';
 import { addReactionToLocalState } from '../../utils/addReactionToLocalState';
@@ -491,6 +491,7 @@ const ChannelWithContext = <
     handleReaction,
     handleRetry,
     handleThreadReply,
+    hasCameraPicker = isImagePickerAvailable(),
     hasCommands = true,
     // If pickDocument isn't available, default to hiding the file picker
     hasFilePicker = pickDocument !== null,
@@ -2225,6 +2226,7 @@ const ChannelWithContext = <
     editMessage,
     emojiSearchIndex,
     FileUploadPreview,
+    hasCameraPicker,
     hasCommands,
     hasFilePicker,
     hasImagePicker,
