+++ package/src/components/Channel/hooks/useCreateInputMessageInputContext.ts
@@ -33,6 +33,7 @@ export const useCreateInputMessageInputContext = <
   editMessage,
   emojiSearchIndex,
   FileUploadPreview,
+  hasCameraPicker,
   hasCommands,
   hasFilePicker,
   hasImagePicker,
@@ -101,6 +102,7 @@ export const useCreateInputMessageInputContext = <
       editMessage,
       emojiSearchIndex,
       FileUploadPreview,
+      hasCameraPicker,
       hasCommands,
       hasFilePicker,
       hasImagePicker,
