+++ package/src/components/MessageInput/InputButtons.tsx
@@ -25,6 +25,7 @@ export type InputButtonsWithContextProps<
   | 'AttachButton'
   | 'CommandsButton'
   | 'giphyActive'
+  | 'hasCameraPicker'
   | 'hasCommands'
   | 'hasFilePicker'
   | 'hasImagePicker'
@@ -46,6 +47,7 @@ export const InputButtonsWithContext = <
     AttachButton,
     CommandsButton,
     giphyActive,
+    hasCameraPicker,
     hasCommands,
     hasFilePicker,
     hasImagePicker,
@@ -69,11 +71,11 @@ export const InputButtonsWithContext = <
     return null;
   }
 
-  return !showMoreOptions && (hasImagePicker || hasFilePicker) && hasCommands ? (
+  return !showMoreOptions && (hasCameraPicker || hasImagePicker || hasFilePicker) && hasCommands ? (
     <MoreOptionsButton handleOnPress={() => setShowMoreOptions(true)} />
   ) : (
     <>
-      {(hasImagePicker || hasFilePicker) && ownCapabilities.uploadFile && (
+      {(hasCameraPicker || hasImagePicker || hasFilePicker) && ownCapabilities.uploadFile && (
         <View
           style={[hasCommands ? styles.attachButtonContainer : undefined, attachButtonContainer]}
         >
@@ -95,6 +97,7 @@ const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = Default
 ) => {
   const {
     giphyActive: prevGiphyActive,
+    hasCameraPicker: prevHasCameraPicker,
     hasCommands: prevHasCommands,
     hasFilePicker: prevHasFilePicker,
     hasImagePicker: prevHasImagePicker,
@@ -105,6 +108,7 @@ const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = Default
 
   const {
     giphyActive: nextGiphyActive,
+    hasCameraPicker: nextHasCameraPicker,
     hasCommands: nextHasCommands,
     hasFilePicker: nextHasFilePicker,
     hasImagePicker: nextHasImagePicker,
@@ -112,6 +116,10 @@ const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = Default
     showMoreOptions: nextShowMoreOptions,
     text: nextText,
   } = nextProps;
+
+  if (prevHasCameraPicker !== nextHasCameraPicker) {
+    return false;
+  }
 
   if (prevHasImagePicker !== nextHasImagePicker) {
     return false;
@@ -158,6 +166,7 @@ export const InputButtons = <
     AttachButton,
     CommandsButton,
     giphyActive,
+    hasCameraPicker,
     hasCommands,
     hasFilePicker,
     hasImagePicker,
@@ -176,6 +185,7 @@ export const InputButtons = <
         AttachButton,
         CommandsButton,
         giphyActive,
+        hasCameraPicker,
         hasCommands,
         hasFilePicker,
         hasImagePicker,
