+++ package/src/contexts/messageInputContext/hooks/useCreateMessageInputContext.ts
@@ -39,6 +39,7 @@ export const useCreateMessageInputContext = <
   FileUploadPreview,
   fileUploads,
   giphyActive,
+  hasCameraPicker,
   hasCommands,
   hasFilePicker,
   hasImagePicker,
@@ -158,6 +159,7 @@ export const useCreateMessageInputContext = <
       FileUploadPreview,
       fileUploads,
       giphyActive,
+      hasCameraPicker,
       hasCommands,
       hasFilePicker,
       hasImagePicker,
